Remove plugin icon
