+++ qgisscripts_plugin/__init__.py
@@ -1,0 +1,3 @@
+def classFactory(iface):
+    from .plugin import ScriptsPlugin
+    return ScriptsPlugin(iface)
